Extract read_var_int into ReadExt trait in read_ext.rs
Move the variable-length integer reader out of event.rs into a new
src/read_ext.rs module as a ReadExt trait, mirroring the ReadBytesExt
pattern. All call sites updated to use r.read_var_int()? syntax.

Also fix tone_groupno argument type (usize, matching v_groupno field)
and its call site cast in service.rs.

+++ src/service.rs
@@ -927,7 +927,7 @@ impl PxtoneService {
         self.units[u].tone_portament(v);
       }
       EVENTKIND_VOICENO => self.moo_reset_voice_on(u, ev.value as usize),
-      EVENTKIND_GROUPNO => self.units[u].tone_groupno(ev.value as u32),
+      EVENTKIND_GROUPNO => self.units[u].tone_groupno(ev.value as usize),
       EVENTKIND_TUNING => self.units[u].tone_tuning(f32::from_bits(ev.value as u32)),
       _ => {} // BEATCLOCK, BEATTEMPO, BEATNUM, REPEAT, LAST は無視
     }
